backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.0.3
psycopg2-binary==2.9.9

# AI Models
openai==1.3.5
anthropic==0.7.8
httpx>=0.24.0,<0.25.0

# Agno Agent Framework  
phidata[openai,anthropic,postgres]>=2.7.0

# Content processing
readability-lxml>=0.8.1
beautifulsoup4==4.12.2
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
numpy>=1.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Evaluation
arize-phoenix>=3.0.0

